+++ CHANGELOG.md
@@ -41,9 +41,11 @@ The git repositories are hosted at the following sites:
 
 ### Added
 
+* Added Who-Is-Router-To-Network process in basic Notification Class when
+  recipient address router MAC address is unknown. (#1243)
 * Add initialization and unpacking functions for BACnet Character
   String buffer structure. (#1242)
-* Add Host-N-Port minimal encode and decode which utilizes the octet
+* Added Host-N-Port minimal encode and decode which utilizes the octet
   string and character string buffer encode and decode. (#1239)
 * Added API for extending the basic Device object and children with
   proprietary properties for ReadProperty and WriteProperty services. (#1238)